Replaced uint32_t with size_t.
User buffer size is not limited by msgpack-c so size_t is the
appropriate type.

+++ include/msgpack/object.h
@@ -99,7 +99,7 @@ MSGPACK_DLLEXPORT
 void msgpack_object_print(FILE* out, msgpack_object o);
 
 MSGPACK_DLLEXPORT
-int msgpack_object_print_buffer(char *buffer, uint32_t buffer_size, msgpack_object o);
+int msgpack_object_print_buffer(char *buffer, size_t buffer_size, msgpack_object o);
 
 MSGPACK_DLLEXPORT
 bool msgpack_object_equal(const msgpack_object x, const msgpack_object y);
